Add support for lograge
lograge enforces a log file format that the standard output for
Tablexi.logger would break. When lograge is detected, and configured in
such a way that it would conflict with our standard logging, we use a
null logger rather than Rails.logger .

Presumably this would only happen in cases where 3rd party error
tracking is enabled, so it wouldn't be a problem to swallow these
messages.

Closes #2